+++ crates/ruma-client-api/src/push/get_pushrule_enabled.rs
@@ -1,36 +1,32 @@
-//! `GET /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/enabled`
+//! `GET /_matrix/client/*/pushrules/global/{kind}/{ruleId}/enabled`
 //!
 //! This endpoint gets whether the specified push rule is enabled.
 
 pub mod v3 {
     //! `/v3/` ([spec])
     //!
-    //! [spec]: https://spec.matrix.org/latest/client-server-api/#get_matrixclientv3pushrulesscopekindruleidenabled
+    //! [spec]: https://spec.matrix.org/latest/client-server-api/#get_matrixclientv3pushrulesglobalkindruleidenabled
 
     use ruma_common::{
         api::{request, response, Metadata},
         metadata,
     };
 
-    use crate::push::{RuleKind, RuleScope};
+    use crate::push::RuleKind;
 
     const METADATA: Metadata = metadata! {
         method: GET,
         rate_limited: false,
         authentication: AccessToken,
         history: {
-            1.0 => "/_matrix/client/r0/pushrules/:scope/:kind/:rule_id/enabled",
-            1.1 => "/_matrix/client/v3/pushrules/:scope/:kind/:rule_id/enabled",
+            1.0 => "/_matrix/client/r0/pushrules/global/:kind/:rule_id/enabled",
+            1.1 => "/_matrix/client/v3/pushrules/global/:kind/:rule_id/enabled",
         }
     };
 
     /// Request type for the `get_pushrule_enabled` endpoint.
     #[request(error = crate::Error)]
     pub struct Request {
-        /// The scope to fetch a rule from.
-        #[ruma_api(path)]
-        pub scope: RuleScope,
-
         /// The kind of rule
         #[ruma_api(path)]
         pub kind: RuleKind,
@@ -48,9 +44,9 @@ pub mod v3 {
     }
 
     impl Request {
-        /// Creates a new `Request` with the given scope, rule kind and rule ID.
-        pub fn new(scope: RuleScope, kind: RuleKind, rule_id: String) -> Self {
-            Self { scope, kind, rule_id }
+        /// Creates a new `Request` with the given rule kind and rule ID.
+        pub fn new(kind: RuleKind, rule_id: String) -> Self {
+            Self { kind, rule_id }
         }
     }