+++ crates/ruma-client-api/src/push/set_pushrule_actions.rs
@@ -1,4 +1,4 @@
-//! `PUT /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/actions`
+//! `PUT /_matrix/client/*/pushrules/global/{kind}/{ruleId}/actions`
 //!
 //! This endpoint allows clients to change the actions of a push rule. This can be used to change
 //! the actions of builtin rules.
@@ -6,7 +6,7 @@
 pub mod v3 {
     //! `/v3/` ([spec])
     //!
-    //! [spec]: https://spec.matrix.org/latest/client-server-api/#put_matrixclientv3pushrulesscopekindruleidactions
+    //! [spec]: https://spec.matrix.org/latest/client-server-api/#put_matrixclientv3pushrulesglobalkindruleidactions
 
     use ruma_common::{
         api::{request, response, Metadata},
@@ -14,25 +14,21 @@ pub mod v3 {
         push::Action,
     };
 
-    use crate::push::{RuleKind, RuleScope};
+    use crate::push::RuleKind;
 
     const METADATA: Metadata = metadata! {
         method: PUT,
         rate_limited: false,
         authentication: AccessToken,
         history: {
-            1.0 => "/_matrix/client/r0/pushrules/:scope/:kind/:rule_id/actions",
-            1.1 => "/_matrix/client/v3/pushrules/:scope/:kind/:rule_id/actions",
+            1.0 => "/_matrix/client/r0/pushrules/global/:kind/:rule_id/actions",
+            1.1 => "/_matrix/client/v3/pushrules/global/:kind/:rule_id/actions",
         }
     };
 
     /// Request type for the `set_pushrule_actions` endpoint.
     #[request(error = crate::Error)]
     pub struct Request {
-        /// The scope to fetch a rule from.
-        #[ruma_api(path)]
-        pub scope: RuleScope,
-
         /// The kind of rule
         #[ruma_api(path)]
         pub kind: RuleKind,
@@ -51,14 +47,9 @@ pub mod v3 {
     pub struct Response {}
 
     impl Request {
-        /// Creates a new `Request` with the given scope, rule kind, rule ID and actions.
-        pub fn new(
-            scope: RuleScope,
-            kind: RuleKind,
-            rule_id: String,
-            actions: Vec<Action>,
-        ) -> Self {
-            Self { scope, kind, rule_id, actions }
+        /// Creates a new `Request` with the given rule kind, rule ID and actions.
+        pub fn new(kind: RuleKind, rule_id: String, actions: Vec<Action>) -> Self {
+            Self { kind, rule_id, actions }
         }
     }